+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Zen Changelog
 
-## [master] - 
+## [4.1] - 2020-03-31
 * Add onboarding configuration
 * Remove author box from single posts
 * Improve blockquotes and pullquotes style